+++ src/client.js
@@ -1,4 +1,3 @@
-
 import Promise from 'bluebird';
 import * as helpers from './helpers';
 
@@ -17,7 +16,7 @@ import TableCompiler from './schema/tablecompiler';
 import ColumnBuilder from './schema/columnbuilder';
 import ColumnCompiler from './schema/columncompiler';
 
-import Pool2 from 'pool2';
+import { Pool } from 'generic-pool';
 import inherits from 'inherits';
 import { EventEmitter } from 'events';
 
@@ -26,6 +25,12 @@ import { assign, uniqueId, cloneDeep } from 'lodash'
 
 const debug = require('debug')('knex:client')
 const debugQuery = require('debug')('knex:query')
+const debugPool = require('debug')('knex:pool')
+
+let id = 0
+function clientId() {
+  return `client${id++}`
+}
 
 // The base client provides the general structure
 // for a dialect specific client object.
@@ -35,6 +40,7 @@ function Client(config = {}) {
   if (this.driverName && config.connection) {
     this.initializeDriver()
     if (!config.pool || (config.pool && config.pool.max !== 0)) {
+      this.__cid = clientId()
       this.initializePool(config)
     }
   }
@@ -152,21 +158,19 @@ assign(Client.prototype, {
     }
   },
 
-  initializePool(config) {
-    if (this.pool) this.destroy()
-    this.pool = new Pool2(assign(this.poolDefaults(config.pool || {}), config.pool))
-    this.pool.on('warn', function(msg) {
-      helpers.warn(`Pool2 - ${msg}`)
-    })
-  },
-
   poolDefaults(poolConfig) {
-    const client = this
+    const name = this.dialect + ':' + this.driverName + ':' + this.__cid
     return {
       min: 2,
       max: 10,
-      acquire(callback) {
-        client.acquireRawConnection()
+      name: name,
+      log(str, level) {
+        if (level === 'info') {
+          debugPool(level.toUpperCase() + ' pool ' + name + ' - ' + str)
+        }
+      },
+      create: (callback) => {
+        this.acquireRawConnection()
           .tap(function(connection) {
             connection.__knexUid = uniqueId('__knexUid')
             if (poolConfig.afterCreate) {
@@ -175,32 +179,48 @@ assign(Client.prototype, {
           })
           .asCallback(callback)
       },
-      dispose(connection, callback) {
+      destroy: (connection) => {
         if (poolConfig.beforeDestroy) {
-          poolConfig.beforeDestroy(connection, function() {
-            if (connection !== undefined) {
-              client.destroyRawConnection(connection, callback)
-            }
-          })
-        } else if (connection !== void 0) {
-          client.destroyRawConnection(connection, callback)
+          helpers.warn(`
+            beforeDestroy is deprecated, please open an issue if you use this
+            to discuss alternative apis
+          `)
+          poolConfig.beforeDestroy(connection, function() {})
+        }
+        if (connection !== void 0) {
+          this.destroyRawConnection(connection)
         }
       },
-      ping(resource, callback) {
-        return client.ping(resource, callback);
+      validate: (connection) => {
+        if (connection.__knex__disposed) {
+          helpers.warn(`Connection Error: ${connection.__knex__disposed}`)
+          return false
+        }
+        return this.validateConnection(connection)
       }
     }
   },
 
+  initializePool(config) {
+    if (this.pool) {
+      helpers.warn('The pool has already been initialized')
+      return
+    }
+    this.pool = new Pool(assign(this.poolDefaults(config.pool || {}), config.pool))
+  },
+
+  validateConnection(connection) {
+    return true
+  },
+
   // Acquire a connection from the pool.
   acquireConnection() {
-    const client = this
     let request = null
-    const completed = new Promise(function(resolver, rejecter) {
-      if (!client.pool) {
-        return rejecter(new Error('There is no pool defined on the current client'))
+    const completed = new Promise((resolver, rejecter) => {
+      if (!this.pool) {
+        return rejecter(new Error('Unable to acquire a connection'))
       }
-      request = client.pool.acquire(function(err, connection) {
+      request = this.pool.acquire(function(err, connection) {
         if (err) return rejecter(err)
         debug('acquired connection from pool: %s', connection.__knexUid)
         resolver(connection)
@@ -220,24 +240,27 @@ assign(Client.prototype, {
   // Releases a connection back to the connection pool,
   // returning a promise resolved when the connection is released.
   releaseConnection(connection) {
-    const { pool } = this
-    return new Promise(function(resolver) {
+    return new Promise((resolver) => {
       debug('releasing connection to pool: %s', connection.__knexUid)
-      pool.release(connection)
+      this.pool.release(connection)
       resolver()
     })
   },
 
   // Destroy the current connection pool for the client.
   destroy(callback) {
-    const client = this
-    const promise = new Promise(function(resolver) {
-      if (!client.pool) return resolver()
-      client.pool.end(function() {
-        client.pool = undefined
-        resolver()
+    const promise = new Promise((resolver) => {
+      if (!this.pool) {
+        return resolver()
+      }
+      this.pool.drain(() => {
+        this.pool.destroyAllNow(() => {
+          this.pool = undefined
+          resolver()
+        })
       })
     })
+
     // Allow either a callback or promise interface for destruction.
     if (typeof callback === 'function') {
       promise.asCallback(callback)