+++ src/dialects/sqlite3/index.js
@@ -59,10 +59,11 @@ assign(Client_SQLite3.prototype, {
 
   // Get a raw connection from the database, returning a promise with the connection object.
   acquireRawConnection() {
-    const client = this;
-    return new Promise(function(resolve, reject) {
-      const db = new client.driver.Database(client.connectionSettings.filename, function(err) {
-        if (err) return reject(err)
+    return new Promise((resolve, reject) => {
+      const db = new this.driver.Database(this.connectionSettings.filename, (err) => {
+        if (err) {
+          return reject(err)
+        }
         resolve(db)
       })
     })
@@ -70,9 +71,12 @@ assign(Client_SQLite3.prototype, {
 
   // Used to explicitly close a connection, called internally by the pool when
   // a connection times out or the pool is shutdown.
-  destroyRawConnection(connection, cb) {
-    connection.close()
-    cb()
+  destroyRawConnection(connection) {
+    connection.close((err) => {
+      if (err) {
+        this.emit('error', err)
+      }
+    })
   },
 
   // Runs the query on the specified connection, providing the bindings and any
@@ -146,13 +150,9 @@ assign(Client_SQLite3.prototype, {
 
   poolDefaults(config) {
     return assign(Client.prototype.poolDefaults.call(this, config), {
-      min: 1,
+      min: 0,
       max: 1
     })
-  },
-
-  ping(resource, callback) {
-    resource.each('SELECT 1', callback);
   }
 
 })