+++ test/integration/index.js
@@ -7,9 +7,11 @@ var logger = require('./logger');
 var config = require('../knexfile');
 var fs     = require('fs');
 
-Object.keys(config).forEach(function(dialectName) {
-  require('./suite')(logger(knex(config[dialectName])));
-});
+var Promise = require('bluebird')
+
+Promise.each(Object.keys(config), function(dialectName) {
+  return require('./suite')(logger(knex(config[dialectName])));
+})
 
 after(function(done) {
   if (config.sqlite3 && config.sqlite3.connection.filename !== ':memory:') {